+++ app/(route)/edit/profile/[id]/_components/ImageEdit.tsx
@@ -3,7 +3,7 @@
 import { FaCameraRetro } from '@react-icons/all-files/fa/FaCameraRetro';
 import { useRef, useState } from 'react';
 
-import { Loader } from '@/app/_components/Loading/Loader';
+import { Loader } from '@/app/_components/ss/Loader';
 import stylse from '@/styles/Images.module.css';
 
 export default function ImageEdit({
