+++ app/_components/ss/skeleton/SkeletonCard.tsx
@@ -1,0 +1,3 @@
+export const SkeletonCard = () => {
+  return <div></div>;
+};
